requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
yfinance>=0.2.28
scikit-learn>=1.3.0
xgboost>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
finnhub-python>=2.4.0